fix compilation error due to ambiguity

+++ src/2geom/toys/sb2d.cpp
@@ -51,7 +51,7 @@ public:
                (2*(iv+vi)/(2.*vi+1)+1)*width/4.);
                double dl = dot((hand.pts[corner+4*i] - base), dir)/dot(dir,dir);
                display_handles[corner+4*i] = dl*dir + base;
-               sb2[i][corner] = dl*10/(width/2)*pow(4.,ui+vi);
+               sb2[i][corner] = dl*10/(width/2)*pow(4.,(double)ui+vi);
            }
         cairo_sb2d(cr, sb2, dir*0.1, width);
